pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-lyceum-fastapi"
version = "1.0.0"
description = "API FastAPI para sincronização com Lyceum"
readme = "README.md"
authors = [
    {name = "Desenvolvedor", email = "dev@example.com"}
]
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.30.0",
    "alembic>=1.12.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=5.9.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.15.1",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "httpx>=0.25.0",
    "aiosqlite>=0.19.0",
]

# ------------------------------------------------------------
# CONFIGURAÇÃO DE DESCOBERTA DE PACOTES (ESCOLHA APENAS ESTA)
# ------------------------------------------------------------
[tool.setuptools.packages.find]
include = ["app*"]          # Inclui todos os pacotes que começam com "app"
exclude = [
    "docker*",
    "migrations*",
    "tests*",
    "scripts*",
    "*.egg-info",
]

# ------------------------------------------------------------
# FERRAMENTAS DE DESENVOLVIMENTO
# ------------------------------------------------------------
[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true